removed unused files and fixed styling after update
