+++ gui/package.json
@@ -12,7 +12,7 @@
   "private": true,
   "dependencies": {
     "@angular/animations": "~9.1.2",
-    "@angular/cdk": "^9.0.0",
+    "@angular/cdk": "^9.2.1",
     "@angular/common": "~9.1.2",
     "@angular/compiler": "~9.1.2",
     "@angular/core": "~9.1.2",
