--- gui/src/app/@core/mock/README.md
@@ -1,1 +1,0 @@
-Application-wise data providers.
