--- gui/src/app/@core/module-import-guard.ts
@@ -1,5 +1,0 @@
-export function throwIfAlreadyLoaded(parentModule: any, moduleName: string) {
-  if (parentModule) {
-    throw new Error(`${moduleName} has already been loaded. Import Core modules in the AppModule only.`);
-  }
-}
