--- gui/src/app/@core/utils/index.ts
@@ -1,1 +1,0 @@
-export { AnalyticsService } from './analytics.service';
